% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_round}
\alias{my_round}
\title{Round with trailing zeroes}
\usage{
my_round(x, digits)
}
\arguments{
\item{x}{Value to round}

\item{digits}{Number of digits to keep}
}
\description{
Round with trailing zeroes
}
