% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tau_CI}
\alias{tau_CI}
\title{Calculate CI for tau from metafor}
\usage{
tau_CI(meta, z.to.r = FALSE)
}
\arguments{
\item{meta}{Object from metafor}

\item{z.to.r}{Should we convert from Fisher's z to r?}
}
\description{
Calculate CI for tau from metafor
}
