% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{z_pvals}
\alias{z_pvals}
\title{Calculate Wald p-values for lmer}
\usage{
z_pvals(.m)
}
\arguments{
\item{.m}{The merMod object from lmer}
}
\description{
These will agree exactly with confint( , method="Wald" ), unlike
lmerTest's too-conservative Satterthwaite approximation.
}
